Verify the binary

+++ setup-gh.sh
@@ -15,13 +15,16 @@ function main () {
   declare -r TOOL_VERSION=${1:-}
   declare -r TOOL_PATH=${2:-${AGENT_TOOLSDIRECTORY:-${RUNNER_TOOL_CACHE:-/tmp/gh}}}
   declare -r DOWNLOAD_PATH=${RUNNER_TEMP:-/tmp}
-  
+
+
   # If the script is running on GitHub, allow debug logging
   # Otherwise, all output is logged
-  if [ ! -z "${GITHUB_ENV:-}" ]; then
+  if [ ! -z "${CI:-}" ]; then
     declare -r DEBUG_CMD=::debug::
+    declare -r ERR_CMD=::error::
   else
     declare -r DEBUG_CMD=''
+    declare -r ERR_CMD='ERROR: '
   fi
 
   # If the version is not specified, identify the latest version
@@ -32,6 +35,18 @@ function main () {
     echo "GH - latest version is v${VERSION}"
   else
     declare -r VERSION=${TOOL_VERSION}
+  fi
+
+  # Just in case it's already available in a global path
+  # outside of the proper tools folder
+  declare -r EXISTING_GH_PATH=$(which gh || echo "")
+  if [ -n "${EXISTING_GH_PATH}" ]; then
+    echo "${DEBUG_CMD}Found existing GitHub CLI installation at ${EXISTING_GH_PATH}"
+    declare -r EXISTING_VERSION=$(${EXISTING_GH_PATH} --version | grep "gh version ${VERSION}")
+    if [ -n "${EXISTING_VERSION}" ]; then
+      echo "${DEBUG_CMD}Existing GitHub CLI installation matches requested version ${VERSION}"
+      exit 0
+    fi
   fi
 
   # Determine the architecture of the system. This assumes
@@ -51,12 +66,45 @@ function main () {
   # If the tool is not already installed, download and install it
   if [ ! -f "${INSTALL_MARKER}" ]; then
     echo "${DEBUG_CMD}Installing GH CLI v${VERSION}"
-    declare -r DOWNLOAD_URL="https://github.com/cli/cli/releases/download/v${VERSION}/gh_${VERSION}_linux_${PLATFORM}.tar.gz"
-    declare -r BINARY_ARCHIVE="cli.tar.gz"
+    declare -r BINARY_ARCHIVE="gh_${VERSION}_linux_${PLATFORM}.tar.gz"
+    declare -r DOWNLOAD_URL="https://github.com/cli/cli/releases/download/v${VERSION}/${BINARY_ARCHIVE}"
+    declare -r RELEASE_SHA=$(curl -SsL https://github.com/cli/cli/releases/download/v${VERSION}/gh_${VERSION}_checksums.txt)
     
+    mkdir -p "${DOWNLOAD_PATH}"
     curl -sLfo "${DOWNLOAD_PATH}/${BINARY_ARCHIVE}" "${DOWNLOAD_URL}"
+
+    pushd "${DOWNLOAD_PATH}" > /dev/null
+
+    if command -v shasum >/dev/null; then
+      if ! echo "${RELEASE_SHA}" | shasum -c --ignore-missing --status; then
+        echo "${ERR_CMD}Checksum verification failed. Aborting installation." >&2
+        exit 1
+      fi
+    else
+      if command -v sha256sum >/dev/null; then
+        if ! echo "${RELEASE_SHA}" | sha256sum -c --ignore-missing --status; then
+          echo "${ERR_CMD}Checksum verification failed. Aborting installation." >&2
+          exit 1
+        fi
+      else
+        echo "${ERR_CMD}Could not verify checksum. Requires sha256sum or shasum command."
+      fi
+    fi
+
+    popd >/dev/null
+    echo "${DEBUG_CMD}Checksum verification passed"
+        
+    # Use attestation to verify the binary and its build provenance
+    # However, this relies on having 'gh' already installed to perform the verification
+    # to avoid simply trusting the results from the downloaded binary.
+    if ! ${EXISTING_GH_PATH} at verify -R cli/cli "${DOWNLOAD_PATH}/${BINARY_ARCHIVE}"; then
+      echo "${ERR_CMD}Attestation verification failed. Aborting installation." >&2
+      exit 1
+    fi
+
     mkdir -p "${INSTALL_PATH}"
     tar xzvf "${DOWNLOAD_PATH}/${BINARY_ARCHIVE}" -C "${INSTALL_PATH}"  --strip-components=1 > /dev/null
+
     rm "${DOWNLOAD_PATH}/${BINARY_ARCHIVE}"
     touch ${INSTALL_MARKER}
   else
